Commit_1
There are still no functions in this version. This is the very initial code where we attempt to test the loop to populate the p-value matrix. We already include two different regressions: one for the first step and one for all the following steps, but we keep getting an error.